refactor: Rename a dir name
