+++ tools/shared/archive_test.go
@@ -1,4 +1,4 @@
-package archive
+package shared
 
 import (
 	"compress/gzip"
